pyproject.toml
[tool.poetry]
name = "fasttext-lite"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "fasttext_lite", from = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
scikit-learn = "^1.2"
numpy = "^1.23"
fasttext-wheel = "^0.9.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.261"
black = "^23.3.0"
pytest = "^7.3.0"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
